Some thread functions
SVN r4117

+++ runtime/gc/thread.h
@@ -13,11 +13,11 @@ typedef struct GC_thread {
    * Furthermore, the exnStack field must be first, because the native
    * codegen depends on this (which is bad and should be fixed).
    */
-  uint32_t exnStack;     /* An offset added to stackBottom that specifies 
-                          * where the top of the exnStack is.
-                          */
-  uint32_t bytesNeeded;  /* The number of bytes needed when returning
-                          * to this thread.
-                          */
-  objptr stack;          /* The stack for this thread. */
+  uint32_t exnStack;  /* An offset added to stackBottom that specifies 
+                       * where the top of the exnStack is.
+                       */
+  size_t bytesNeeded; /* The number of bytes needed when returning
+                       * to this thread.
+                       */
+  objptr stack;       /* The stack for this thread. */
 } *GC_thread;
